requirements.txt
streamlit
pandas
matplotlib
openpyxl
arabic-reshaper
python-bidi